Limit the recirc_id to max of 2^^28 if tc hw-offload is enabled.
netdev-offload-tc when offloading a flow to tc, uses the flow's
recirc_id as flower chain id.  recirc_id is of type 'uint32_t'.
Kernel tc reserves the upper nibble (4 bits) of the tc flower's
chain index for extended action opcode [1].  If a flow's recirc_id
value is greater than 268,435,455 (0x0fffffff), kernel returns
EINVAL when ovs-vswitchd tries to offload the flow.

This commit fixes this offload issue by setting the maximum
value of recirc id to 268,435,455 when hw-offload is enabled.

[1] - https://elixir.bootlin.com/linux/v6.17.6/source/tools/include/uapi/linux/pkt_cls.h#L50-L64

Suggested-by: Ilya Maximets <i.maximetes@ovn.org>
Signed-off-by: Numan Siddique <numans@ovn.org>