ofproto-dpif: Limit recirc_id to 28 bits for TC compatibility.
The netdev-offload-tc module uses a flow's recirc_id (a uint32_t)
as the TC flower chain ID during offloading.  However, the kernel's
TC implementation reserves the upper 4 bits of the chain index for
extended action opcodes.  If a recirc_id exceeds (2^28 - 1), the
kernel rejects the flow offload with EINVAL.  This patch caps the
maximum recirc_id to (2^28 - 1) to prevent these failures.

This limit is sufficiently high that it is unlikely to impact the
number of active datapath flows in practice.

Reported-at: https://mail.openvswitch.org/pipermail/ovs-dev/2025-November/427485.html
Suggested-by: Ilya Maximets <i.maximets@ovn.org>
Suggested-by: Eelco Chaudron <echaudro@redhat.com>
Acked-by: Eelco Chaudron <echaudro@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>